updated for missing type hints (PDO) in functions fetchAllDB and deleteBook and corrected for PSR12 violations. Added is_numeric to if statment before calling deleteBook function

+++ index.php
@@ -4,9 +4,9 @@ require_once "functions.php";
 
 $db = connectDB();
 
-if (isset($_POST['delete'])) {
+if (isset($_POST['delete']) && is_numeric($_POST['delete'])) {
   $deleteID = deleteBook($db, $_POST['delete']);
-}
+  }
 
 $books = fetchAllDB($db);
 ?>
